updated tests with some cleanup and splitting file per class
