+++ test/Microsoft.Azure.SignalR.IntegrationTests/MessageOrderTests.cs
@@ -7,20 +7,19 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Xunit;
+using Xunit.Abstractions;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.AspNetCore.SignalR.Protocol;
-using Microsoft.Azure.SignalR.IntegrationTests.Infrastructure;
-using Microsoft.Azure.SignalR.IntegrationTests.MockService;
 using Microsoft.Azure.SignalR.Protocol;
 using Microsoft.Azure.SignalR.Tests.Common;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-
-using Xunit;
-using Xunit.Abstractions;
-
+using Microsoft.Azure.SignalR.IntegrationTests.Infrastructure;
+using Microsoft.Azure.SignalR.IntegrationTests.Infrastructure.MessageOrderTests;
+using Microsoft.Azure.SignalR.IntegrationTests.MockService;
 using AspNetTestServer = Microsoft.AspNetCore.TestHost.TestServer;
 
 namespace Microsoft.Azure.SignalR.IntegrationTests
