+++ test/Microsoft.Azure.SignalR.IntegrationTests/Microsoft.Azure.SignalR.IntegrationTests.csproj
@@ -8,6 +8,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Infrastructure\MessageOrderTestHubs\**" />
+    <EmbeddedResource Remove="Infrastructure\MessageOrderTestHubs\**" />
+    <None Remove="Infrastructure\MessageOrderTestHubs\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\Microsoft.Azure.SignalR.Tests.Common\Microsoft.Azure.SignalR.Tests.Common.csproj" />
   </ItemGroup>
